test(e2e): show merged feature flags in Allure report overview (QAA-1360)
chore(e2e): drop comments from merged-FF Allure change (QAA-1360)
